Enable cancellation of input prompts
This change enables the cancellation of input prompts by calling a
function rather than using Ctrl+C in the console.  This is necessary in
cases like stopping the debugger where an active input prompt can cause
the debugger to hang if it can't be cancelled.  The fix is to be more
reactive to the abort of command execution so that the active input
prompt can be cancelled.

Resolves #428.